Fix the tests and fixtures
