+++ mypy/test/data/check-modules.test
@@ -409,12 +409,12 @@ __all__.append('c')
 __all__.extend(('d', 'e'))
 
 a = b = c = d = e = f = _g = 1
-[builtins fixtures/module.py]
+[builtins fixtures/module_all.py]
 
 [case testAllMustBeSequenceStr]
 import typing
 __all__ = [1, 2, 3]
-[builtins fixtures/module.py]
+[builtins fixtures/module_all.py]
 [out]
 main: error: Type of __all__ must be Sequence[str], not List[int]
