+++ mypy/test/data/fixtures/module.py
@@ -1,6 +1,3 @@
-from typing import Generic, Sequence, TypeVar
-_T = TypeVar('_T')
-
 class object:
     def __init__(self) -> None: pass
 class module: pass
@@ -8,8 +5,3 @@ class type: pass
 class function: pass
 class int: pass
 class str: pass
-class list(Generic[_T], Sequence[_T]):
-    def append(self, x: _T): pass
-    def extend(self, x: Sequence[_T]): pass
-    def __add__(self, rhs: Sequence[_T]) -> list[_T]: pass
-class tuple: pass
